// apiCalls/meals/dto/response/get_history.dto.ts
export interface GetHistoryDto {
  _id: string;
  image: string;
  donorId: string;
  foodDesc: string;
  veg: boolean;
  feedsUpto: number;
  address: string;
  city: string;
  state: string;
  country: string;
  postalCode: number;
  preferredTime: string;
  deliveryDate: string | null;
  expiryDate: string;
  status: "delivered" | "expired" | "cancelled";
  collectorId: string | null;
  createdAt: string;
  updatedAt: string;
}

export type GetHistoryResponseDto = GetHistoryDto[];